rf: move healthcheck to own class
