+++ lib/api/BackbeatAPI.js
@@ -6,10 +6,8 @@ const zookeeper = require('node-zookeeper-client');
 const { errors } = require('arsenal');
 
 const BackbeatProducer = require('../BackbeatProducer');
+const Healthcheck = require('./Healthcheck');
 const routes = require('./routes');
-
-// In-Sync Replicas
-const ISRS = 3;
 
 /**
  * Class representing Backbeat API endpoints and internals
@@ -35,6 +33,7 @@ class BackbeatAPI {
         this._crrProducer = null;
         this._crrStatusProducer = null;
         this._metricProducer = null;
+        this._healthcheck = null;
         this._zkClient = null;
     }
 
@@ -61,33 +60,13 @@ class BackbeatAPI {
             && this._crrStatusProducer.isReady();
     }
 
-    _getConnectionDetails() {
-        return {
-            zookeeper: {
-                status: this._zkClient.getState().name === 'SYNC_CONNECTED' ?
-                    'ok' : 'error',
-                details: this._zkClient.getState(),
-            },
-            kafkaProducer: {
-                status: this._checkProducersReady() ? 'ok' : 'error',
-            },
-        };
-    }
-
     /**
-     * Checks health of in-sync replicas
-     * @param {object} md - topic metadata object
-     * @return {boolean} true if ISR health is ok
+     * Check if Zookeeper and Producer are connected
+     * @return {boolean} true/false
      */
-    _checkISRHealth(md) {
-        // eslint-disable-next-line consistent-return
-        const keys = Object.keys(md);
-        for (let i = 0; i < keys.length; i++) {
-            if (md[keys[i]].isr && md[keys[i]].isr.length !== ISRS) {
-                return 'error';
-            }
-        }
-        return 'ok';
+    isConnected() {
+        return this._zkClient.getState().name === 'SYNC_CONNECTED'
+            && this._checkProducersReady();
     }
 
     /**
@@ -96,40 +75,22 @@ class BackbeatAPI {
      * @return {undefined}
      */
     healthcheck(cb) {
-        const client = this._crrProducer.getKafkaClient();
-
-        client.loadMetadataForTopics([], (err, res) => {
+        return this._healthcheck.getHealthcheck((err, data) => {
             if (err) {
-                this._logger.trace('error getting healthcheck');
+                this._logger.error('error getting healthcheck', {
+                    error: err.message,
+                });
                 return cb(errors.InternalError);
             }
-            const response = res.map(i => (Object.assign({}, i)));
-            const connections = {};
-            try {
-                const topicMD = {};
-                response.forEach((obj, idx) => {
-                    if (obj.metadata && obj.metadata[this._repConfig.topic]) {
-                        const copy = JSON.parse(JSON.stringify(obj.metadata[
-                            this._repConfig.topic]));
-                        topicMD.metadata = copy;
-                        response.splice(idx, 1);
-                    }
-                });
-                response.push(topicMD);
-
-                connections.isrHealth = this._checkISRHealth(topicMD.metadata);
-            } finally {
-                Object.assign(connections, this._getConnectionDetails());
-
-                response.push({
-                    internalConnections: connections,
-                });
-
-                return cb(null, response);
-            }
+            return cb(null, data);
         });
     }
 
+    /**
+     * Setup internals
+     * @param {function} cb - callback(error)
+     * @return {undefined}
+     */
     setupInternals(cb) {
         async.parallel([
             done => this._setZookeeper(done),
@@ -159,6 +120,9 @@ class BackbeatAPI {
                 this._logger.error('error setting up internal clients');
                 return cb(err);
             }
+            this._healthcheck = new Healthcheck(this._repConfig, this._zkClient,
+                this._crrProducer, this._crrStatusProducer,
+                this._metricProducer);
             this._logger.info('BackbeatAPI setup ready');
             return cb();
         });
